Api Major Change v2.0
Updated the API with new responsibilities, how it works and how it responds. The API itself should reply much faster, and even to multiple requests at the same time.

Some bugs around right now, comes from the database itself. There need to be updated the collation charset, for how data is stored on the database. But also fixed how the API output some of the data, so it's camelcase instead of pascal.