Bot Changes - v2.2.0
Improved requests of handling events and running the process of the clan application.
Removed the enums class, and separated the enums into specific classes, without a class. But just a namespace, for easier access. Reworked the program file, to only containing the events being listed on, so the clan application process handles what it needs, no further.