[Bot Changes] - v2.7.0
Removed objects / models, which are now globally accessible through the Data Library.
Changed interfaces to point to the right objects and models.

Feature:
Working with Command Modules.
Setting up logging and information, so it's possible to get a list of latest logs and necessary information.
Added Dependency Injection for the Minion.