[Database Update]
Have changed all objects, removed regions for each object, would hide the information more that it was being useful. Added a few new objects to the V3 API folder, so it takes usage of the newest changes. Added support for admins to the endpoint call on either Clan or Clan by id.

Repositories has been changed to the newest change also with a bit cleaner code, which is easier to understand.